Added interface.js and modified data objects.
Added a new Interface file that will be in charge of creating the
UI of the application. index.js will be in charge of adding events
to classes and perform application logic.

Modified styles.css to have the option of a class called completed,
which will put a line through text (used for todos that are marked
as complete).

Modified project/todo.js files that added a static method to return
a private array of project/todos in their respected files.