Added dblclick event to projects.
Fixed bug where if a user clicked a project, an error will show
in the console saying the project is invalid. The fix to this is
to check if the selection is null, and if true, return.

Added dblclick to index.js, which does nothing right now, but in
the next commit, will make a project name editable by dblclicking
the project name and pressing the enter key.